+++ cmd/cbc/main.go
@@ -6,7 +6,7 @@ import (
 )
 
 type flags struct {
-   bandwidth int
+   bandwidth int64
    email string
    id string
    mech.Stream
@@ -21,7 +21,7 @@ func main() {
    // e
    flag.StringVar(&f.email, "e", "", "email")
    // f
-   flag.IntVar(&f.bandwidth, "f", 2052370, "video bandwidth")
+   flag.Int64Var(&f.bandwidth, "f", 2052370, "video bandwidth")
    // g
    flag.StringVar(&f.name, "g", "English", "audio name")
    // i
