+++ cmd/paramount/main.go
@@ -10,7 +10,7 @@ import (
 )
 
 type flags struct {
-   bandwidth int
+   bandwidth int64
    codecs string
    dash bool
    guid string
@@ -33,7 +33,7 @@ func main() {
    // d
    flag.BoolVar(&f.dash, "d", false, "DASH download")
    // f
-   flag.IntVar(&f.bandwidth, "f", 1611000, "video bandwidth")
+   flag.Int64Var(&f.bandwidth, "f", 1611000, "video bandwidth")
    // g
    flag.StringVar(&f.codecs, "g", "mp4a", "audio codec")
    // h
